Initialize TestExceptionCallback in new threads
In a KJ_TEST, it is expected that when we get an ERROR-severity log
message, the entire test program should exit with a non-zero return
code.  This allows us to catch ERRORs in CI.

In a test that I wrote using threads, I noticed my test program still
succeeded (by exiting with a 0 return code) when the program logged an
ERROR-severity log message in the non-main thread.  This happens
because TestExceptionCallback does not implement getThreadInitializer,
so we get a default RootExceptionCallback.  The RootExceptionCallback
will write errors to stderr, but doesn’t inform the ProcessContext
that an error has occured and the process should exit with a non-zero
return code.

The way to fix this is to implement getThreadInitializer in
TestExceptionCallback.  My implementation is simple: in a new thread,
I construct a new TestExceptionCallback and then run the thread’s main
function.

TestExceptionCallback now takes the `ProcessContext` as a `const`
reference to mark that the `ProcessContext` may be shared between
threads and that we should only call thread-safe (`const`) methods.

To test that this code works, I applied the following patch:

```
modified   c++/src/kj/test-test.c++
@@ -19,8 +19,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.

+#include "async.h"
+#include "async-unix.h"
 #include "common.h"
 #include "test.h"
+#include "thread.h"
 #include <cstdlib>
 #include <stdexcept>
 #include <signal.h>
@@ -51,6 +54,18 @@ KJ_TEST("expect signal from sigint") {
   KJ_EXPECT_SIGNAL(kj::none, raise(SIGINT));
 }

+KJ_TEST("expect failure from thread log") {
+  UnixEventPort port1;
+  EventLoop loop1(port1);
+  WaitScope waitScope1(loop1);
+
+  PromiseCrossThreadFulfillerPair<void> xpaf = kj::newPromiseAndCrossThreadFulfiller<void>();
+  kj::Thread thread([&]() noexcept {
+    KJ_LOG(ERROR, "ERROR log message from thread");
+    xpaf.fulfiller->fulfill();
+  });
+}
+
 }  // namespace
 }  // namespace _
 }  // namespace kj
```

When run with this commit, the “expect failure from thread log” test
reports as PASS, but the test program fails.