Mark ProcessContext’s `warning` and `error` methods as const
I plan to call these methods from multiple threads, so KJ’s style
guide says they must be marked as const.

The only implementation of `ProcessContext` that I’m aware of is
`TopLevelProcessContext`.  The `warning` and `error` methods of
`TopLevelProcessContext` should be thread safe because `writeLineToFd`
writes to stack-local data and a file descriptor and `hadError` is an
atomic.

Changing the constness of the methods is a backwards-incompatible
change, so other implementations will be broken until they implement
the const methods.