Propagate test failures into parent TestExceptionCallbacks
If there’s an exception in a child thread, we’d like the entire test
to fail.  In order to do this, we must tell the top-most
TestExceptionCallback that a test has failed.  I do this by
propagating a reference to the main thread’s TestExceptionCallback to
all spawned threads.

I tested this with the following patch:

```
modified   c++/src/kj/test-test.c++
@@ -19,8 +19,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.

+#include "async.h"
+#include "async-unix.h"
 #include "common.h"
 #include "test.h"
+#include "thread.h"
 #include <cstdlib>
 #include <stdexcept>
 #include <signal.h>
@@ -51,6 +54,18 @@ KJ_TEST("expect signal from sigint") {
   KJ_EXPECT_SIGNAL(kj::none, raise(SIGINT));
 }

+KJ_TEST("expect failure from thread log") {
+  UnixEventPort port1;
+  EventLoop loop1(port1);
+  WaitScope waitScope1(loop1);
+
+  PromiseCrossThreadFulfillerPair<void> xpaf = kj::newPromiseAndCrossThreadFulfiller<void>();
+  kj::Thread thread([&]() noexcept {
+    KJ_EXPECT(false);
+    xpaf.fulfiller->fulfill();
+  });
+}
+
 }  // namespace
 }  // namespace _
 }  // namespace kj
```

which causes the test to FAIL.