Create API for RpcSystem to notify VatNetwork when connections are idle.
See comments in rpc.h for an explanation.

This change also has a subtle effect on behavior even when the hint is ignored: When the RpcSystem is destroyed, or when it receives a FIN from the other side, previously it would always send an error message to the other side. Now, it will only do so if the connection was not idle. If the connection is idle, then nothing is really lost, and we can just send FIN.