Improve hashCode().
Previously, `hashCode()` merely needed to return a value that was likely to be unique. When the input was already a 32-bit integer, it was acceptable to just return exactly the input value. However, this meant that the distribution of hashes wasn't necessarily uniform. To compensate for this, the hashtable implementation would use prime-number bucket counts. This, however, made computing the correct bucket expensive, as it required an integer division operation.

After this change, `hashCode()` is expected to return a uniform distribution. Thus, hashtables can have power-of-two size, avoiding the integer division. This improves performance in our string-keyed hashtable benchmark by 10%. The integer-keyed benchmark is slower, due to the need to spend more cycles computing hashes, but integer keys are already very fast anyway and most real-world hash tables tend to be string-keyed.

This also improves the behavior of open addressing with linear probing (which KJ's hashtable uses). Previously, if you inserted many sequential keys into a hash table, they would occupy consecutive buckets, creating a large range of buckets that were full. If you then tried to insert a non-sequential key that happened to hash somewhere into the range of occupied buckets, the hashtable wolud have to scan all the way to the end of the range to find an appropriate slot for the new key. Cap'n Proto's RPC tables tended to have this problem in particular.