Drive-by: Bring workerd's stack symbolization into KJ.
This mostly just copies the code from workerd directly int kj/exception.c++:
https://github.com/cloudflare/workerd/blob/2ee31a40c5f46b912b4f4d8940bb5a03f86bd4e2/src/workerd/util/symbolizer.c++

This makes KJ once again produce reasonable stack trace symbolication (with source line numbers) by default, even with ASLR and even with dynamic linking. Historically, KJ's original symbolication only worked with fixed-position executables (which was the default at the time it was first written, but hasn't been for a long time).

This only works when linking with libdl. Perhaps we should make that a requirement for v2.