Fix http-over-capnp server-side promise ordering issue.
The server side calls `clientRequestContext.startResponse()` to send the response headers. If the response has no body, then `HttpService::request()` could immediately return here. Since we've `exclusiveJoin()`ed the `startResponse()` promise with `HttpService::request()`, the `startResponse()` call is actually canceled as a result.

This hasn't shown up as a problem because promise ordering was such that the `startResponse()` call managed to actually get sent before being canceled. However, when I did some stuff to mess with the timing, I saw this break.

Instead, we `detach()` the RPC. It's ugly but in this case it's safe.