WebSocketImpl::disconnect() should now throw.
As of the recent signature change for WebSocket::disconnect(), we now say that it should be safe to call in destructors (which you often wan to do, because it's used to signal that the outgoing side of the WebSocket was dropped).

WebSocketImpl::disconnect(), however, still contains an assertion of `!currentlySending`. I had thought this assertion made sense because calling disconnect() is UB if a send is in progress. However, I forgot that `currentlySending` remains true even after a send is canceled, at which point `disconnect()` should actually be safe to call. This assertion should be fine to just remove.