Simplify CapnpToKjWebSocketAdapter::cancel().
This method was doing more than it needed to. See comments for explanation.

In particular, the call to `disconnect()` was triggering the bug in the previous commit and throwing an exception, in which case we didn't end up nulling out the refs, leading to UAF later on.

Now that the `disconnect()` bug is fixed, we could in theory leave `cancel()` as-is, but this change should make it more robust against other lurking bugs.