rpc-test: Rename: TestNetworkAdapter -> TestVat
The naming here was really confusing, especially as there was also `TestNetwork` and instances of both `TestNetwork` and `TestNetworkAdapter` tended to be named "network".

I used `TestVat` instead of `TestVatNetwork` because in the test we have multiple of these representing different vats, so calling each one a "network" is weird. (In the real world, you'd have only one `VatNetwork` because your program is only one vat, so the "network" part of the name makes more sense there -- it's your view into the network.)

(I suspect that when this test was first created, `VatNetwork` was called `NetworkAdapter`, hence the naming here.)