Fix stringifyStackTraceWithLlvm().
The code was double-closing the stdout pipe on the parent's side, due to the fact that fdopen() takes ownership of the fd, but the code was also closing the fd itself. Errors from close() were being ignored so the problem wasn't noticed. However, when running in a multi-threaded context, the code would occasionally race with an FD being created in another thread, and end up closing that FD which it didn't own. This manifested as a non-deterministic "bad file descriptor" error in srs-test in the workers runtime.

I rewrote the code to use RAII and to check all errors (catching them with an outer try/catch).