Fix bug in in-memory AsyncCapabilityStream implementation.
I noticed this bug while converting things to KJ_SYSCALL_FD. Decided to call out the bug in a separate commit and add a test.

An almost-identical code block on lines 1024-1026 don't have the same bug. Both code blocks go back to #971 when this functionality was introduced. Neither one was covered by tests before this commit. Not sure why I had apparently fixed one but not the other.