Add Executor::newPromiseAndCrossThreadFulfiller()
Currently, `newPromiseAndCrossThreadFulfiller()` can only create Promises associated with the calling thread's event loop. This is the overwhelmingly common use case for CrossThreadPromiseFulfillers.

Recently, while working on the ability to `co_await` Rust Futures across an FFI boundary, I encountered a case where it would be helpful to create Promises associated with a different thread's event loop.

To elaborate on the use case:
- Rust Futures have a `poll()` function which drives the Future to make some forward progress.
- The caller of `poll()` must pass a reference to a Waker object to `poll()`. Let's call this Waker the "initial Waker". It lives on the stack of the thread which calls `poll()`.
- Waker is an abstract interface, with any number of actual implementations.
- If the Future is not yet ready, it stores a clone of the initial Waker somewhere, and arranges to call `Waker::wake()` asynchronously on the clone when it thinks it can next make forward progress. Let's call this second Waker the "cloned Waker".
- Rust may call `Waker::wake()` on the cloned Waker from any thread.
- So, when awaiting Futures from KJ, CrossThreadPromiseFulfiller serves as a good implementation for the cloned Waker. The associated Promise can be arranged to fire an Event which calls `Future::poll()`.
- But creating a CrossThreadPromiseFulfiller has some overhead -- memory allocations, and a slow-path toward fulfilling the Promise. So ideally, we won't create the CrossThreadPromiseFulfiller unless we know we actually need it -- that is, when Rust tries to clone the initial Waker.
- The only problem: a Rust Future may clone the Waker from a different thread, as long as it does so before the call to `poll()` on the main thread returns. For this reason, it would be helpful to be able to call `newPromiseAndCrossThreadFulfiller()` from any thread.

This commit adds an `Executor::newPromiseAndCrossThreadFulfiller()` member function, which allows for the above use case. It's a very narrow use case, and unlikely to be commonly useful. But, the Rust API allows it, so I need to handle it.

This change exposed a footgun: callers needed to arrange to move the promise back to its native event loop before that event loop exited, or else they risk UB. To make this situation safer, the XThreadPaf now holds a strong reference to the Executor, and fulfilling such a Promise after its event loop has exited is now a no-op.