RPC bugfix: Properly catch exceptions from ClientHook::call().
Arguably, ClientHooks should not throw from this -- they should return a broken promise. But there are a lot of ClientHook implementations and they could really be doing anything.

Previously, an exception from here would propagate right out of `handleCall()` and kill the connection read loop, aborting the connection. Another side effect is that the `RpcCallContext` constructed in `handleCall()` would be destroyed without having sent a response, which would cause it to send a "canceled" return, despite the client not having actually asked to cancel anything. We are seeing these unexpected "canceled" returns in the wild.