more powerful as<T>() implementation (#2396)
Current design converts `as<T>()`` calls into static members of T, which
doesn't make it extensible: T needs to know about every type that can
be converted.

Updating the design to convert as<T>() in `into(T{}, this)`, which
makes T a simple marker object and you can have as many conversion targets
as you like.