Revert "Try to debug connections being marked idle while still having references."
This reverts commit e7b011a2c1c028bd89d82a536e737757b2bc7ee4.

The logs seem to all look like one of the following:

```
refcounts = 1, 1, 0, 0, 1, 0, 0, 0
refcounts = 1, 2, 0, 0, 2, 0, 0, 0
refcounts = 1, 1, 0, 0, 0, 0, 0, 0
refcounts = 1, 2, 0, 0, 0, 0, 0, 0
refcounts = 1, 0, 0, 0, 1, 0, 0, 0
```

refcounts[0] (CONNECTION_MAP) is expected to be 1, and the rest are expected to be 0. So refcounts[1] (CLIENT) and [4] (PIPELINE) are unexpectedly non-zero.

On closer inspection, these are easily explained: `RpcPipeline` and `PromiseClient` (subclass of `RpcClient`) can both point at capabilities that end up resolving to a local object, thus no longer requiring the RPC connection and no longer holding an entry in any of the tables. But, as written, they will still hold onto the `RpcConnectionState`.

Arguably, these classes should be refactored to drop their references when no longer needed. But, that would be a tricky refactor. In any case, though, these references do not actually need to block idleness, so the next commit will change the logic to avoid this.