Fix "Connection shut down due to being idle".
It turns out looking at the refcount doesn't work because some objects keep refcounts on the connection even when they no longer actually need it and therefore do not need to block idle closure.

Instead of checking the refcount, we should focus on verifying that we do not need to call disconnect(), because all the tables it would service are empty.