Increase default WebSocket max message size to 32MiB
Downstream in workerd, we increased the WebSocket maximum message size
to 32MiB, by adjusting the value we pass to WebSocket::receive(maxSize).
Unfortunately, there are some other places where WebSocket::receive() is
called, such as in the WebSocket::pumpTo() fallback implementation, and
there is currently no good way to plumb the necessary parameter all the
way down -- it'll require an API change in kj-http.

Instead of shaving the yak, we're just increasing the default maximum
message size to match what we publicly committed to downstream. In the
future, we can make this limit actually configurable.