kj/exception: KJ_USE_BACKTRACE to configure backtraces at compile
KJ_USE_BACKTRACE can enable (1) or disable (0) backtraces regardless
of platform where previously it was decided by platform.

Previously KJ_HAS_BACKTRACE could be set to 1 to enable if auto
detection was failing for some reason, but disabling it was not
respected.

We maintain the backwards compat of KJ_HAS_BACKTRACE and the
default platform detection if neither is defined.

See #1080.