More aggressively use "depth-first" promise ordering.
This means: When running an event causes another event to be scheduled, we should put that new event at the front of the queue, so it runs next, before anything else that had been scheduled before this event ran.

(If one event schedules N new events, we put all those events at the beginning, but we run them in the order they were scheduled.)

This approach means the event loop stays focused on one task until there is nothing left to do, before switching to a diffrent task. This is good, because:
* Less switching between tasks is nicer to the CPU cache.
* It tends to reduce latency. If N identical tasks run with regular switching between them, they will all complete in about the same amount of time -- the time to complete all the tasks. But if we focus on one task to completion before moving on to the next, then the first task will take 1/N as much time, the next 2/N, etc., with the last task taking just as much time as when interleaving. Overall, average latency is halved.
* Behavior is more predictable and race conditions less likely as unrelated tasks will not jump in and change the state of the world in the middle of a task.

There is a possible objection: If we always run a task to completion before moving on, what if it doesn't complete? A task might go into an infinite loop. In this case, the other tasks in the queue will be starved.

This objection is naive: The KJ event loop is a cooperative scheduler. It is not designed to handle scenarios in which the CPU is saturated, regardless of scheduling order. Even with strictly FIFO ordering, one task scheduling event endlessly will ruin performance for everyone else. Such heavyweight tasks should instead be scheduled to a separate thread.